+++ src/main/java/cn/memoryzy/json/service/persistent/converter/BlacklistConverter.java
@@ -2,7 +2,7 @@ package cn.memoryzy.json.service.persistent.converter;
 
 import cn.hutool.core.codec.Base64;
 import cn.hutool.core.util.StrUtil;
-import cn.memoryzy.json.model.JsonEntry;
+import cn.memoryzy.json.model.BlacklistEntry;
 import cn.memoryzy.json.model.wrapper.ArrayWrapper;
 import cn.memoryzy.json.model.wrapper.ObjectWrapper;
 import cn.memoryzy.json.util.Json5Util;
@@ -17,16 +17,16 @@ import java.util.LinkedList;
  * @author Memory
  * @since 2025/3/3
  */
-public class BlacklistConverter extends Converter<LinkedList<JsonEntry>> {
+public class BlacklistConverter extends Converter<LinkedList<BlacklistEntry>> {
 
     @Override
-    public @Nullable LinkedList<JsonEntry> fromString(@NotNull String value) {
-        LinkedList<JsonEntry> jsonEntries = new LinkedList<>();
+    public @Nullable LinkedList<BlacklistEntry> fromString(@NotNull String value) {
+        LinkedList<BlacklistEntry> jsonEntries = new LinkedList<>();
         value = StrUtil.str(Base64.decode(value), StandardCharsets.UTF_8);
         ArrayWrapper jsonArray = Json5Util.parseArray(value);
         for (Object data : jsonArray) {
             ObjectWrapper element = (ObjectWrapper) data;
-            JsonEntry entry = JsonEntry.fromMap(element);
+            BlacklistEntry entry = BlacklistEntry.fromMap(element);
             jsonEntries.add(entry);
         }
 
@@ -34,7 +34,7 @@ public class BlacklistConverter extends Converter<LinkedList<JsonEntry>> {
     }
 
     @Override
-    public @Nullable String toString(@NotNull LinkedList<JsonEntry> value) {
+    public @Nullable String toString(@NotNull LinkedList<BlacklistEntry> value) {
         return Base64.encode(Json5Util.compressJson5(value));
     }
 }