+++ src/main/java/cn/memoryzy/json/ui/listener/ListRightClickPopupMenuMouseAdapter.java
@@ -1,7 +1,5 @@
 package cn.memoryzy.json.ui.listener;
 
-import cn.memoryzy.json.model.JsonEntry;
-
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.MouseAdapter;
@@ -14,10 +12,10 @@ import java.util.Objects;
  */
 public class ListRightClickPopupMenuMouseAdapter extends MouseAdapter {
 
-    private final JList<JsonEntry> showList;
+    private final JList<?> showList;
     private final JPopupMenu popupMenu;
 
-    public ListRightClickPopupMenuMouseAdapter(JList<JsonEntry> showList, JPopupMenu popupMenu) {
+    public ListRightClickPopupMenuMouseAdapter(JList<?> showList, JPopupMenu popupMenu) {
         this.showList = showList;
         this.popupMenu = popupMenu;
     }
@@ -28,7 +26,7 @@ public class ListRightClickPopupMenuMouseAdapter extends MouseAdapter {
             int x = e.getX();
             int y = e.getY();
 
-            JsonEntry selectedValue = showList.getSelectedValue();
+            Object selectedValue = showList.getSelectedValue();
             if (Objects.isNull(selectedValue)) {
                 int index = showList.locationToIndex(new Point(x, y));
                 if (index != -1) {